162 profile menu (#163)
* fix layout

* fix formatting

* fix title

* fix individual company rendering

* create form template

* fix edit company route

* fix layout

* edit route

* fix button

* fix company create

* Split profile editing into new templates

* create file for later splitting of routes

* rename

* remove endif

* change buttons to links to new templates

* add routes for profile and password edit templates

* remove unneeded kwargs

* fix routes

* edit edit_password route

* delete unused templates

* split profiles and companies routes

* fix comment

* Move into subfolders

* fix url

* fix include

* fix url

* remove comment

* remove value attributes

* remove comments
