+++ app/routes/__init__.py
@@ -3,5 +3,5 @@ from app.routes.auth import auth as auth_blueprint
 from app.routes.forums import forums_blueprint
 from app.routes.blogPosts import blogPosts as blogPosts_blueprint
 from app.routes.profiles import profiles as profile_blueprint
-from app.routes.profiles import company_info as company_blueprint
+from app.routes.companies import company_info as company_blueprint
 from app.routes.tickets import ticket_blueprint
