+++ app/routes/auth.py
@@ -10,7 +10,7 @@ auth = Blueprint('auth', __name__)
 
 @auth.route('/login')
 def login():
-    return render_template('login.html')
+    return render_template('authentication/login.html')
 
 @auth.route('/login', methods=['POST'])
 def login_post():
@@ -27,7 +27,7 @@ def login_post():
 
 @auth.route('/signup')
 def signup():
-    return render_template('signup.html')
+    return render_template('authentication/signup.html')
 
 @auth.route('/signup', methods=['POST'])
 def signup_post():
