+++ app/routes/main.py
@@ -6,12 +6,12 @@ main = Blueprint('main', __name__)
 @main.route("/")
 @main.route("/home")
 def home():
-    return render_template('home.html')
+    return render_template('home/home.html')
 
 @main.route("/about")
 def about():
-    return render_template('about.html', title ='about')
+    return render_template('about/about.html', title ='about')
 
 @main.route("/products")
 def products():
-    return render_template('products.html', title ='products')
+    return render_template('products/products.html', title ='products')
