+++ app/routes/profiles.py
@@ -1,84 +1,49 @@
-from flask import Blueprint, render_template, redirect, url_for, request
+from flask import Blueprint, render_template, redirect, url_for, request,flash
 from flask_login import login_required, current_user
+from werkzeug.security import generate_password_hash, check_password_hash
 from app.db import db
-from app.db.models import User, Company, CompanyMembers
-
+from app.db.models import User
 
 profiles = Blueprint('profiles', __name__, url_prefix="/profile")
-company_info = Blueprint('company_info', __name__, url_prefix="/companies")
 
 @profiles.route('/')
 @login_required
 def profile():
     return render_template('profile/profile.html', name=current_user.name, email=current_user.email)
 
-@profiles.route('/edit_profile', methods=['POST'])
+@profiles.route('/edit_profile', methods=['GET','POST'])
 @login_required
 def edit_profile_post():
-    old_email = request.form.get('old_email')
-    name = request.form.get('name')
-    email = request.form.get('email')
-    user = User.query.filter_by(email=old_email).first()
-    user.name = name
-    user.email = email
-    db.session.commit()
-    return redirect(url_for('profiles.profile'))
-
-@profiles.route('/edit_password')
-def edit_password():
-    return render_template('edit_password.html')
-
-@profiles.route('/edit_password', methods=['POST'])
-def edit_password_post():
-    return redirect(url_for('profiles.profile'))
-
-def get_companies():
-    return Company.query.all()
-
-@company_info.route("/")
-def display_companies_home():
-    return render_template('company_info/company_info_main.html', companies=get_companies())
-
-@company_info.route('/<company_id>')
-@login_required
-def display_company_info(company_id):
-    company = Company.query.get_or_404(company_id)
-    return render_template('company_info/company_info.html', company=company)
-
-@company_info.route('/create', methods=['GET', 'POST'])
-@login_required
-def create_company():
     if request.method == 'POST':
+        old_email = request.form.get('old_email')
         name = request.form.get('name')
-        address1 = request.form.get('address1')
-        address2 = request.form.get('address2')
-        city = request.form.get('city')
-        state = request.form.get('state')
-        zipcode = request.form.get('zipcode')
-        country = request.form.get('country')
-        phone = request.form.get('phone')
         email = request.form.get('email')
-        new_company = Company(name=name, address1=address1, address2=address2, city=city, state=state, zipcode=zipcode, country=country, phone=phone, email=email)
-        db.session.add(new_company)
+        user = User.query.filter_by(email=old_email).first()
+        user.name = name
+        user.email = email
         db.session.commit()
-        new_members_company = CompanyMembers(id = new_company.id , user_id = current_user.id)
-        db.session.add(new_members_company)
-        db.session.commit()
-        return redirect(url_for("company_info.display_company_info", company_id=new_company.id))
-    return render_template('company_info/company_info_create.html', companies=get_companies())
+        return redirect(url_for('profiles.profile'))
 
-@company_info.route('/edit_company_info/<company_id>', methods=['POST'])
-@login_required
-def edit_company_info_post(company_id):
-    company = Company.query.filter_by(id=company_id).first()
-    company.name = request.form.get('name')
-    company.address1 = request.form.get('address1')
-    company.address2 = request.form.get('address2')
-    company.city = request.form.get('city')
-    company.state = request.form.get('state')
-    company.zipcode = request.form.get('zipcode')
-    company.country = request.form.get('country')
-    company.phone = request.form.get('phone')
-    company.email = request.form.get('email')
-    db.session.commit()
-    return redirect(url_for('company_info.display_company_info', company_id=company.id))
+    return render_template('profile/profile_edit.html', name=current_user.name, email=current_user.email)
+
+@profiles.route('/edit_password', methods=['GET','POST'])
+def edit_password():
+    if request.method == 'POST':
+        email = request.form.get('email')
+        old_password = request.form.get('curr_password')
+        new_password = request.form.get('new_password')
+        confirm_new_password = request.form.get('confirm_new_password')
+        user = User.query.filter_by(email=email).first()
+
+        if new_password == confirm_new_password:
+            if check_password_hash(user.password, old_password):
+                user.password = generate_password_hash(new_password, method='sha256')
+                db.session.add(user)
+                db.session.commit()
+
+                return redirect(url_for('profiles.profile'))
+
+        flash('Please check your password details and try again.') # does nothing fix later
+
+    return render_template('profile/password_edit.html', name=current_user.name, email=current_user.email)
+
