+++ app/templates/about/about.html
@@ -1,14 +1,14 @@
-      
-{% extends "oneDiv.html" %}
+{% extends "layout/oneDiv.html" %}
 {% block oneDivContent %}
     <!--Flask uses jinja2 a templating engine. lets you write code in the template-->
     <!--for loop to loop through the list of posts -->
     <article class="media content-section">
         <div class="media-body">
           <div class="article-metadata">
-            <a class="mr-2" href="#">Pengin Open Source</a>
+            <a class="mr-2" href="#">About</a>
+            <small class="text-muted">July 14th, 2022</small>
           </div>
-          <h2><a class="article-title" href="#">Welcome to Pengin Open Source</a></h2>
+          <h2><a class="article-title" href="#">Pengin Open Source</a></h2>
           <p class="article-content">Building cool tech for the future.</p>
         </div>
       </article>
