+++ app/templates/authentication/signup.html
@@ -1,4 +1,4 @@
-{% extends "oneDiv.html" %}
+{% extends "layout/oneDiv.html" %}
 
 {% block oneDivContent %}
 <div class="column md-6 offset-3">
