+++ app/templates/blog/blog_main.html
@@ -1,4 +1,4 @@
-{% extends "twoAndOneThirdDivs.html" %}
+{% extends "layout/twoAndOneThirdDivs.html" %}
 
 {% block twoThirdContent %}
     <div class="blog-content">
