--- app/templates/blog_main.html
@@ -1,22 +1,0 @@
-{% extends "layout.html" %}
-
-{% block content %}
-    <div class="blogContent">
-        {% block blogContent %}{% endblock %}
-    </div>
-{% endblock %}
-
-{% block rightblock %}
-    <h3>links</h3>
-    {% for link in links %}
-        <li>
-            <a
-            class="font-weight-bold text-info"
-            href="/blog/{{ link.id }}">
-                {{ link.title }}
-            </a>
-        </li>
-    {% endfor %}
-{% endblock %}
-
-
