+++ app/templates/company_info/company_edit.html
@@ -1,0 +1,66 @@
+{% extends "profile/profile_main.html" %}
+
+
+{% block profilecontent %}
+<form id="edit_company_info_form" method="POST" action="/companies/edit_company_info/{{company.id}}">
+
+  <h1 class="header">Edit Company Information</h1>
+
+  <input type="hidden" name="id" value="{{company.id}}">
+
+  <div class="form-group">
+      <label>Company Name*</label>
+      <input class="form-control" type="text" name="name" value="{{ company.name }}" placeholder="Company Name" required>
+  </div>
+
+  <div class="form-group">
+      <label>Address 1*</label>
+      <input class="form-control" type="text" name="address1" value="{{ company.address1 }}" placeholder="Address 1" required>
+    </div>
+
+  <div class="form-group">
+      <label>Address 2</label>
+      <input class="form-control" type="text" name="address2" value="{{ company.address2 }}" placeholder="Address 2">
+  </div>
+
+  <div class="form-group">
+      <label>City*</label>
+      <input class="form-control" type="text" name="city" value="{{ company.city }}" placeholder="City" required>
+  </div>
+
+  <div class="form-group">
+    <label>State*</label>
+    <input class="form-control" type="text" name="state" value="{{ company.state }}" placeholder="State">
+  </div>
+
+  <div class="form-group">
+      <label>ZIP*</label>
+      <input class="form-control" type="text" name="zipcode" value="{{ company.zipcode }}" placeholder="ZIP" required>
+    </div>
+
+  <div class="form-group">
+      <label>Country*</label>
+      <input class="form-control" type="text" name="country" value="{{ company.country }}" placeholder="Country" required>
+  </div>
+
+  <div class="form-group">
+      <label>Phone</label>
+      <input class="form-control" type="text" name="phone" value="{{ company.phone }}" placeholder="Phone">
+  </div>
+
+  <div class="form-group">
+      <label>Email</label>
+      <input class="form-control" type="text" name="email" value="{{ company.email }}" placeholder="Email">
+  </div>
+
+  <div class="container d-flex justify-content-between px-0">
+    <button class="btn btn-success btn-md" type="submit">Save</button>
+  </div>
+
+</form>
+{% endblock %}
+
+
+{% block profileSideContent %}
+  {% include "right_window/profile_info.html" %}
+{% endblock %}
