--- app/templates/edit_profile.html
@@ -1,39 +1,0 @@
-<!-- 
-  This is the page that users will see to edit their user information. 
--->
-<!--Andy: Not needed anymore, form loads in profile page-->
-
-{% extends "layout.html" %}
-
-{% block content %}
-<div class="column is-10">
-
-<h3 class="title">Edit Profile</h3>
-    <div class="box">
-        <form method="POST" action="/edit_profile">
-
-        <div class="form-group">
-            <label for="name">New Name:</label><br>
-            <input class="form-control" type="text" name="name" placeholder="" autofocus="" value={{name}}>
-        </div>
-
-            <!-- <div class="field">
-                <div class="control">
-                    <label for="email">Email:</label><br>
-                    <input class="input is-large" type="email" name="email" placeholder="" autofocus="">
-                </div>
-            </div> -->
-
-            <!-- <div class="field">
-                <div class="control">
-                    <label for="password">New Password:</label><br>
-                    <input class="input is-medium" type="password" name="password" placeholder="">
-                </div>
-            </div> -->
-
-            <button class="btn btn-primary btn-md">Save Updates</button>
-        </form>
-    </div>
-</div>
-
-{% endblock %}
