+++ app/templates/forums/alternating_color_table.html
@@ -1,4 +1,4 @@
-{% extends "oneDiv.html" %}
+{% extends "layout/oneDiv.html" %}
 
 {% block oneDivContent %}
 {% block beforeTableContent %} {% endblock %}
