+++ app/templates/forums/create_post.html
@@ -1,4 +1,4 @@
-{% extends "oneDiv.html" %}
+{% extends "layout/oneDiv.html" %}
 
 {% block oneDivContent %}
   <div class="column">
