+++ app/templates/layout/midArticleDiv.html
