+++ app/templates/products/products.html
@@ -1,5 +1,5 @@
-      
-{% extends "oneThirdDivs.html" %}
+{% extends "layout/oneThirdDivs.html" %}
+
 {% block oneThirdContentLeft %}
 <article class="media content-section">
     <div class="media-body">
