--- app/templates/profile.html
@@ -1,38 +1,0 @@
-<!-- Logan Kiser: this is no longer called in the view, see profile/profile.html -->
-
-{% extends "layout.html" %}
-
-{% block content %}
-
-<div class="profile-info">
-
-  <form method="GET">
-    <h2>Profile Information: Welcome, {{name}}!</h2>
-      <table class="table table-bordered">
-        <tr>
-          <th width="30%">Name</th>
-          <td width="2%">:</td>
-          <td>{{ name }}</td>
-        </tr>
-        <tr>
-          <th width="30%">Email</th>
-          <td width="2%">:</td>
-          <td>{{ email }}</td>
-        </tr>
-      </table>
-      <div class="buttons">
-        <button class="button is-block is-info is-medium" formaction="/edit_profile">Edit Profile</button>
-        <!-- for testing purposes: -->
-        <button class="button is-block is-info is-medium" formaction="/edit_password">Edit Password</button> 
-      </div>
-  </form>
-
-  {% block profilecontent %}{% endblock %}
-  
-</div>
-  
-{% endblock %}
-
-{% block rightblock %}
-  {% include "right_window/profile_info.html" %}
-{% endblock %}
