+++ app/templates/profile/password_edit.html
@@ -1,0 +1,27 @@
+{% extends "profile/profile_main.html" %}
+
+{% block profilecontent %}
+<form id="edit_password_form"method="POST" action="/profile/edit_password">
+    <h1 class="header">Edit Password</h1>
+    <input type="hidden" name="email" value="{{email}}">
+    <div class="form-group">
+      <label>Current Password</label>
+      <input class="form-control" type="password" name="curr_password" placeholder="Your old password">
+    </div>
+    <div class="form-group">
+      <label>New Password</label>
+      <input class="form-control" type="password" name="new_password"  placeholder="Your new password">
+    </div>
+    <div class="form-group">
+      <label>Confirm New Password</label>
+      <input class="form-control" type="password" name="confirm_new_password" placeholder="Confirm your new password">
+    </div>
+    <div class="container d-flex justify-content-between px-0">
+      <button class="btn btn-success btn-md" type="submit">Save</button>
+    </div>
+  </form>
+{% endblock %}
+
+{% block profileSideContent %}
+  {% include "right_window/profile_info.html" %}
+{% endblock %}
