+++ app/templates/profile/profile_edit.html
@@ -1,0 +1,23 @@
+{% extends "profile/profile_main.html" %}
+
+{% block profilecontent %}
+  <form id="edit_profile_form"  method="POST" action="/profile/edit_profile">
+    <h1 class="header">Edit Profile</h1>
+    <input type="hidden" name="old_email" value="{{email}}">
+    <div class="form-group">
+        <label>Name</label>
+        <input class="form-control" type="name" name="name" value="{{ name }}" placeholder="Your new name">
+    </div>
+    <div class="form-group">
+      <label>Email</label>
+      <input class="form-control" type="email" name="email" value="{{ email }}" placeholder="Your new email">
+    </div>
+    <div class="container d-flex justify-content-between px-0">
+      <button class="btn btn-success btn-md" type="submit">Save</button>
+    </div>
+  </form>
+{% endblock %}
+
+{% block profileSideContent %}
+  {% include "right_window/profile_info.html" %}
+{% endblock %}
