+++ app/templates/right_window/profile_info.html
@@ -1,5 +1,3 @@
-
-<h3>Profile</h3>
 <p class='text-muted'>
     <ul class="list-group">
         <a href="/profile" style="text-decoration:none;"><button type="button" class="list-group-item list-group-item-action">Personal Info</button></a>
