+++ app/templates/tickets/create_ticket.html
@@ -1,4 +1,4 @@
-{% extends "oneDiv.html" %}
+{% extends "layout/oneDiv.html" %}
 
 {% block oneDivContent %}
   <div class="column">
